Populate basic directory structure
Starting at the very beginning, the ‘best practices’ Ansible playbook
directory structure is created. At this level, there are no files yet
and the directories simply serve as an structuring/organizational tool.
In the roles folder, a ‘common’ role is made for all regular Ansible
tasks (such as installing nginx, updating, etc). The security folder is
used for configuring IPtables and sshd to secure the server. Likewise,
the gameserver folder would contain tasks that are used for the team
fortress 2 server.