Moved swap tasks to swap.yml
In this commit, I decided to reorganise the tasks in the optimize role.
Rather than having the add swap and NTP sections all in one big file,
I’ll have a modular system of individual .yml files for each task, and
have all of the tasks included in the main.ml.