Fixed fallocate already existing error
When the script is executed twice on the same machine, the swap file
may already exist. If that happens, fallocate would fail since it’s
trying to create a file that’s already being used by system. I fixed it
by adding the create argument to command. If the file already exists,
Ansible would skip the fallocate task.