Fixed ssh lockout bug
It turns out that a weird jinja2 templating decision made it so that
all variables set as no somehow defaults to the value 'False'. This
caused a misconfiguration in the sshd_config file which resulted in the
file being invalid - and we were able to correct it by changing the
variable values to strings with quotes - so Jinja2 does not think it is
a Boolean Value.